how to execute a simple LINQ query and process the results, how to add, modify, and delete data, and how to build entity data models for an existing database, such as Northwind

+++ Product.cs
@@ -13,7 +13,7 @@ namespace Packt.Shared
         public string ProductName { get; set; }
 
         [Column("UnitPrice", TypeName = "money")]
-        public decimal? Cost { get; set; }
+        public float? Cost { get; set; }
 
         [Column("UnitsInStock")]
         public short? Stock { get; set; }
